cain/requirements.txt
numpy==1.17.0
opencv-python==3.4.2.17
Pillow==9.4.0
tensorboard==2.0.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.39.0
